Epetra, AztecOO: Fixing Trilinos bugs 5791, 5793, 5794, 5797 reported by Karen Devine, all related to 64 bit Epetra.  The biggest issue was converting index base to long long like other quantities (Max GID etc).
Build/Test Cases Summary
Enabled Packages: Epetra, Teuchos, ShyLU, Zoltan2, Amesos2
Disabled Packages: SEACAS,Sundance,PyTrilinos,STK
Enabled all Forward Packages
0) MPI_DEBUG => FAILED: passed=1533,notpassed=7 => Not ready to push! (349.22 min)
1) SERIAL_RELEASE => passed: passed=1491,notpassed=0 (146.50 min)
WARNING: Forced the push!

+++ packages/epetra/src/Epetra_BlockMapData.cpp
@@ -50,7 +50,7 @@
 #define EPETRA_BLOCKMAP_NEW_LID
 
 //=============================================================================
-Epetra_BlockMapData::Epetra_BlockMapData(long long NumGlobalElements, int ElementSize, int IndexBase, const Epetra_Comm & Comm, bool IsLongLong) 
+Epetra_BlockMapData::Epetra_BlockMapData(long long NumGlobalElements, int ElementSize, long long IndexBase, const Epetra_Comm & Comm, bool IsLongLong) 
   : Comm_(Comm.Clone()),
     Directory_(0),
     LID_(0),
