+++ packages/epetra/src/Epetra_Map.h
@@ -143,7 +143,7 @@ class EPETRA_LIB_DLL_EXPORT Epetra_Map : public Epetra_BlockMap {
   Epetra_Map(int NumGlobalElements, int IndexBase, const Epetra_Comm& Comm);
 #endif
 #ifndef EPETRA_NO_64BIT_GLOBAL_INDICES
-  Epetra_Map(long long NumGlobalElements, int IndexBase, const Epetra_Comm& Comm);
+  Epetra_Map(long long NumGlobalElements, long long IndexBase, const Epetra_Comm& Comm);
 #endif
 
 
@@ -179,7 +179,7 @@ class EPETRA_LIB_DLL_EXPORT Epetra_Map : public Epetra_BlockMap {
   Epetra_Map(int NumGlobalElements, int NumMyElements, int IndexBase, const Epetra_Comm& Comm);
 #endif
 #ifndef EPETRA_NO_64BIT_GLOBAL_INDICES
-  Epetra_Map(long long NumGlobalElements, int NumMyElements, int IndexBase, const Epetra_Comm& Comm);
+  Epetra_Map(long long NumGlobalElements, int NumMyElements, long long IndexBase, const Epetra_Comm& Comm);
 #endif
 
 
@@ -227,7 +227,7 @@ class EPETRA_LIB_DLL_EXPORT Epetra_Map : public Epetra_BlockMap {
 #ifndef EPETRA_NO_64BIT_GLOBAL_INDICES
   Epetra_Map(long long NumGlobalElements, int NumMyElements,
              const long long *MyGlobalElements,
-             int IndexBase, const Epetra_Comm& Comm);
+             long long IndexBase, const Epetra_Comm& Comm);
 #endif
   
   //! Epetra_Map copy constructor.
